app/src/test/java/com/futo/platformplayer/BackgroundTaskHandlerTests.kt
package com.futo.platformplayer

import com.futo.platformplayer.constructs.BackgroundTaskHandler
import kotlinx.coroutines.*
import kotlinx.coroutines.test.runTest
import org.junit.Assert.assertEquals
import org.junit.Before
import org.junit.After
import org.junit.Test
import kotlinx.coroutines.test.resetMain
import kotlinx.coroutines.test.setMain
import kotlinx.coroutines.test.StandardTestDispatcher

@OptIn(ExperimentalCoroutinesApi::class)
class BackgroundTaskHandlerTest {

    @Before
    fun setup() {
        Dispatchers.setMain(StandardTestDispatcher())
    }

    @After
    fun cleanup() {
        Dispatchers.resetMain()
    }

    @Test
    fun testRun() = runTest {
        var result = 0
        val taskHandler = BackgroundTaskHandler(this, { result = 1 })

        taskHandler.run()
        delay(1000)

        assertEquals(1, result)
    }

    @Test
    fun testCancel() = runTest {
        var result = 0
        val taskHandler = BackgroundTaskHandler(this, { Thread.sleep(500); result = 1 })

        taskHandler.run()
        taskHandler.cancel()
        delay(1000)

        assertEquals(0, result) // Should still be 0 because task was cancelled
    }

    //TODO: Disabled to fix pipeline for now.
    /*
    @Test
    fun testException() = runTest {
        var exceptionMessage: String? = null
        val taskHandler = BackgroundTaskHandler(this, { throw Exception("Test exception") })

        taskHandler.exception<Exception> { exceptionMessage = it.message }
        taskHandler.run()
        delay(1000)

        assertEquals("Test exception", exceptionMessage)
    }*/

    @Test
    fun testDispose() = runTest {
        var result = 0
        val taskHandler = BackgroundTaskHandler(this, { Thread.sleep(500); result = 1 })

        taskHandler.run()
        taskHandler.dispose()
        delay(1000)

        assertEquals(0, result) // Should still be 0 because task was disposed
    }
}